Comment out if str[i]=="'" and str[i+1]=="'" check
The subject conditional check that prints the letter "a" creates confusing output for the customer and it does not appear this check is needed for other than debugging. Recommending to comment it out. Also updated "offset" print to "Processing offset". If this is accepted, please update the script that is in the Teradata DDL extract zip. Reach out to Mike Gohl for more details on inclusion in the zip.